removed winner logic from board, moved into rules class
